LayoutTests/http/tests/xmlhttprequest/access-control-preflight-sync-method-denied-expected.txt

CONSOLE MESSAGE: line 1: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?state=method. Method DELETE is not allowed by Access-Control-Allow-Methods response header field.
PASS: Request successfully blocked.
